- prefer to use video record pin than video preoview in AugmentedMediaSource - resolve a silly mistake that crahed in release mode in CameraKSPropertyInquiry - reorder UI operations for showing the camera preview that was causing issue - simplifying the SimpleMediaSource ColorMode custom proeprty and renaming structure to avoid confusion - remove unused code in systemTray and WinRT component - some nit renaming and addressing comments

+++ Samples/VirtualCamera/VirtualCameraManager_WinRT/CameraKsPropertyInquiry.h
@@ -12,7 +12,7 @@ namespace winrt::VirtualCameraManager_WinRT::implementation
         static winrt::VirtualCameraManager_WinRT::IExtendedPropertyPayload GetExtendedControl(winrt::VirtualCameraManager_WinRT::ExtendedControlKind const& extendedControlKind, winrt::Windows::Media::Devices::VideoDeviceController const& controller);
         static void SetExtendedControlFlags(winrt::VirtualCameraManager_WinRT::ExtendedControlKind const& extendedControlKind, winrt::Windows::Media::Devices::VideoDeviceController const& controller, uint64_t flags);
         static winrt::VirtualCameraManager_WinRT::ISimpleCustomPropertyPayload GetSimpleCustomControl(winrt::VirtualCameraManager_WinRT::SimpleCustomControlKind const& customControlKind, winrt::Windows::Media::Devices::VideoDeviceController const& controller);
-        static void SetSimpleCustomControlFlags(winrt::VirtualCameraManager_WinRT::SimpleCustomControlKind const& customControlKind, winrt::Windows::Media::Devices::VideoDeviceController const& controller, uint64_t flags);
+        static void SetSimpleCustomControlFlags(winrt::VirtualCameraManager_WinRT::SimpleCustomControlKind const& customControlKind, winrt::Windows::Media::Devices::VideoDeviceController const& controller, uint32_t flags);
         static winrt::VirtualCameraManager_WinRT::IAugmentedMediaSourceCustomPropertyPayload GetAugmentedMediaSourceCustomControl(
             winrt::VirtualCameraManager_WinRT::AugmentedMediaSourceCustomControlKind const& customControlKind, 
             winrt::Windows::Media::Devices::VideoDeviceController const& controller);
